+++ src/cli/mod.rs
@@ -13,6 +13,7 @@ pub mod file_ops;
 pub(crate) mod file_ops_aws;
 pub(crate) mod helpers;
 pub(crate) mod local_ops;
+pub(crate) mod ls_view;
 pub(crate) mod migrate_ops;
 pub(crate) mod scan_ops;
 pub(crate) mod secret_ops;